Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/2dcontext/path-objects/2d.path.arc.selfintersect.1-expected.txt

2d.path.arc.selfintersect.1
arc() with lineWidth > 2*radius is drawn sensibly
Actual output:
Expected output:


PASS arc() with lineWidth > 2*radius is drawn sensibly 
